change: refactoring binding rule
* separate in several files
* make interface
