+++ gemini-binder/src/commonMain/kotlin/com/haroncode/gemini/binder/rule/BindingRule.kt
@@ -1,0 +1,7 @@
+package com.haroncode.gemini.binder.rule
+
+
+interface BindingRule {
+
+    suspend fun bind()
+}
